+++ CHANGELOG.md
@@ -4,10 +4,21 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-# [1.23.3] - 2022-09-13
+# [1.24.0] - 2022-09-21
+
+### Added
+
+- ioTracker - ioTracker3
+
+# [1.23.3] - 2022-09-20
+
+### Added
+
+- Enginko - EGK-LW20W00
 
 ### Changed
 
+- Adnexo - updated meta
 - Xovis - Added plugin data output
 
 # [1.23.2] - 2022-09-08